+++ lib/taro/types/scalar/iso8601_date_type.rb
@@ -2,11 +2,10 @@ class Taro::Types::Scalar::ISO8601DateType < Taro::Types::ScalarType
   self.desc = 'Represents a time as Date in ISO8601 format.'
   self.openapi_name = 'ISO8601Date'
   self.openapi_type = :string
-
-  PATTERN = /\A\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])\z/
+  self.pattern = /\A\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])\z/
 
   def coerce_input
-    if object.instance_of?(String) && object.match?(PATTERN)
+    if object.instance_of?(String) && object.match?(pattern)
       Date.parse(object)
     else
       input_error("must be a ISO8601 formatted string")