+++ lib/taro/types/scalar/iso8601_datetime_type.rb
@@ -2,11 +2,10 @@ class Taro::Types::Scalar::ISO8601DateTimeType < Taro::Types::ScalarType
   self.desc = 'Represents a time as DateTime in ISO8601 format.'
   self.openapi_name = 'ISO8601DateTime'
   self.openapi_type = :string
-
-  PATTERN = /\A\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])T([01]\d|2[0-3]):[0-5]\d:[0-5]\d(Z|[+-](0[0-9]|1[0-4]):[0-5]\d)\z/
+  self.pattern = /\A\d{4}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])T([01]\d|2[0-3]):[0-5]\d:[0-5]\d(Z|[+-](0[0-9]|1[0-4]):[0-5]\d)\z/
 
   def coerce_input
-    if object.instance_of?(String) && object.match?(PATTERN)
+    if object.instance_of?(String) && object.match?(pattern)
       DateTime.iso8601(object)
     else
       input_error("must be a ISO8601 formatted string")