+++ lib/taro/types/scalar_type.rb
@@ -1,5 +1,6 @@
 # Abstract base class for scalar types, i.e. types without fields.
 class Taro::Types::ScalarType < Taro::Types::BaseType
+  include Taro::Types::Shared::Pattern
 end
 
 module Taro::Types::Scalar